dlls/atidlxx: Stub GPU/display enumeration and clock/memory info.
Shadow of War uses atiadlxx to query the following GPU information:

 * current memory frequency
 * current core frequency
 * memory size
 * memory bandwidth

Those are used to apprise the GPU and recommend user settings.

Most of them can be fetched via AMD's powerplay sysfs entries / wined3d
helpers. The only value that is not exposed is the memory bandwith so we
set it to the value my RX580 reports on Windows.

Because of that on high end system the game may recommend lower settings
than it would with the real value, and the other way around on low end
systems.

We also report some "sane default" in case we can't find the powerplay
entries - this may be the case with nvapi hack enabled when we report
any NVIDIA card as RX480.

ADL2 stubs used by Call of Duty games are untouched and still report 0
GPUs.