server: Implement HID rawinput message support.
The messages are of variable size and carry the corresponding HID report
data after each RAWINPUT struct.

For now hDevice is set to 0 but it will need to be a pseudo-handle that
will be used to uniquely identify the device and possibly query its
information using GetRawInputDeviceInfo.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=50506
Signed-off-by: Rémi Bernon <rbernon@codeweavers.com>