hidclass.sys: Send RawInput messages for report_id-less devices correctly.
The HID_XFER_PACKET contains the report verbatim, i.e. report id is included
only if it's used by the device.

WM_INPUT / GetRawInputData() just like HidD_GetInputReport() *always* uses the
first byte for report id.

If IDs are not used by the device - then it is set to 0, which is reserved by
the HID spec.

This fixes Hades and other games using SDL 2.0.14+ for handling xinput
controllers.